requirements.txt
asgiref==3.7.2
black==23.10.0
boto3==1.34.0
botocore==1.34.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==4.2.6
django-allauth==0.58.2
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
jmespath==1.0.1
Markdown==3.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
whitenoise==6.6.0
zipp==3.17.0

# Uncomment these lines to use a Postgres database. Both are needed, since in production
# (which uses Linux) we want to install from source, so that security updates from the
# underlying Heroku stack image are picked up automatically, thanks to dynamic linking.
# On other platforms/in development, the precompiled binary package is used instead, to
# speed up installation and avoid errors from missing libraries/headers.
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"

django-allauth==0.58.2
django-storages==1.14.2
python-dateutil==2.8.2
urllib3==1.26.18
jmespath==1.0.1
boto3==1.34.0
botocore==1.34.0
s3transfer==0.9.0
six==1.16.0